Exercise 1-11
Added some test cases in terms of text files to be piped as input to my_wc to test that it works correctly.
Also added a text explaination of what to test, in file Exercise_1-11.txt.
I need to script the testing to make it automated. I was thinking Python, because I like Python.  Maybe bash or some other shell would be better. More UNIX-y.